+++ src/lib/api.ts
@@ -1,5 +1,5 @@
-
 export { CleanOptions } from './tasks/clean';
+export { CheckRepoActions } from './tasks/check-is-repo';
 export { GitError } from './errors/git-error';
 export { GitResponseError } from './errors/git-response-error';
 export { TaskConfigurationError } from './errors/task-configuration-error';